+++ livechat/configuration/api/v36.py
@@ -15,8 +15,10 @@ class ConfigurationApiV36(HttpClient):
                  base_url: str,
                  http2: bool,
                  proxies=None,
-                 verify: bool = True):
-        super().__init__(token, base_url, http2, proxies, verify)
+                 verify: bool = True,
+                 disable_logging: bool = False):
+        super().__init__(token, base_url, http2, proxies, verify,
+                         disable_logging)
         self.api_url = f'https://{base_url}/v3.6/configuration/action'
 
 # Agents
@@ -1306,9 +1308,9 @@ class ConfigurationApiV36(HttpClient):
         '''
         if payload is None:
             payload = prepare_payload(locals())
-        return self.session.get(f'{self.api_url}/get_product_source',
-                                json=payload,
-                                headers=headers)
+        return self.session.post(f'{self.api_url}/get_product_source',
+                                 json=payload,
+                                 headers=headers)
 
     def reactivate_email(self,
                          agent_id: str = None,
@@ -1328,9 +1330,9 @@ class ConfigurationApiV36(HttpClient):
         '''
         if payload is None:
             payload = prepare_payload(locals())
-        return self.session.get(f'{self.api_url}/reactivate_email',
-                                json=payload,
-                                headers=headers)
+        return self.session.post(f'{self.api_url}/reactivate_email',
+                                 json=payload,
+                                 headers=headers)
 
 
 # Batch requests