+++ livechat/configuration/base.py
@@ -26,7 +26,8 @@ class ConfigurationApi:
         base_url: str = api_url,
         http2: bool = False,
         proxies: dict = None,
-        verify: bool = True
+        verify: bool = True,
+        disable_logging: bool = False,
     ) -> Union[ConfigurationApiV33, ConfigurationApiV34, ConfigurationApiV35]:
         ''' Returns client for specific Configuration API version.
 
@@ -42,6 +43,7 @@ class ConfigurationApi:
                                verify the identity of requested hosts. Either `True` (default CA bundle),
                                a path to an SSL certificate file, an `ssl.SSLContext`, or `False`
                                (which will disable verification). Defaults to `True`.
+                disable_logging (bool): indicates if logging should be disabled.
 
             Returns:
                 ConfigurationApi: API client object for specified version.
@@ -50,14 +52,18 @@ class ConfigurationApi:
                 ValueError: If the specified version does not exist.
         '''
         client = {
-            '3.3': ConfigurationApiV33(token, base_url, http2, proxies,
-                                       verify),
-            '3.4': ConfigurationApiV34(token, base_url, http2, proxies,
-                                       verify),
-            '3.5': ConfigurationApiV35(token, base_url, http2, proxies,
-                                       verify),
-            '3.6': ConfigurationApiV36(token, base_url, http2, proxies,
-                                       verify),
+            '3.3':
+            ConfigurationApiV33(token, base_url, http2, proxies, verify,
+                                disable_logging),
+            '3.4':
+            ConfigurationApiV34(token, base_url, http2, proxies, verify,
+                                disable_logging),
+            '3.5':
+            ConfigurationApiV35(token, base_url, http2, proxies, verify,
+                                disable_logging),
+            '3.6':
+            ConfigurationApiV36(token, base_url, http2, proxies, verify,
+                                disable_logging),
         }.get(version)
         if not client:
             raise ValueError('Provided version does not exist.')